Remove wrong @return array from StatsdDataFactoryInterface
Most methods have either no `@return` or `@return mixed`, which
allows implementations to use a buffer or otherwise have deferred or
void methods. I believe that part is intentional since the internal
objects from StatsdDataInterface as returned via produceStatsdData in
the default implementation are afaik never used by end-users.

The set() and increment() methods had a phpdoc `@return array` which
appears to be there by mistake, as StatsdDataFactory actually does
not return array for those methods.

In Wikimedia's BufferingStatsdDataFactory implementation this led to
warnings from Phan static analysis, such as:

> PhanTypeMissingReturn Method BufferingStatsdDataFactory::set is
> declared to return array in phpdoc but has no return value
> PhanTypeMissingReturn Method BufferingStatsdDataFactory::increment
> is declared to return array in phpdoc but has no return value

Remove these for consistency with timing() and gauge(), and so that
downstream can eventually remove suppressions for these issues.